refactored query locations!
