+++ routes/app/pagination/accountPaymentIntents.tsx
@@ -1,7 +1,7 @@
 // Api endpoints for the pagination API
 import { Handlers } from "$fresh/server.ts";
 import { getPagination, getTotalPages } from "../../../lib/backend/businessLogic.ts";
-import { selectPaymentIntentsByAccountIdPaginated, selectPaymentIntentsByAccountIdPaginatedWithSearch } from "../../../lib/backend/db/pagination.ts";
+import { selectPaymentIntentsByAccountIdPaginated, selectPaymentIntentsByAccountIdPaginatedWithSearch } from "../../../lib/backend/db/tables/PaymentIntents.ts";
 import { errorResponseBuilder } from "../../../lib/backend/responseBuilders.ts";
 import { MapPaymentIntentsTableColNamesToDbColNames, PAYMENTINTENTSPAGESIZE, PaymentIntentsTableColNames } from "../../../lib/enums.ts";
 import { State } from "../../_middleware.ts";
